Pull checktime tests out of api_tests into its own file. Also moved some test_utils code out of unittests into chain_test_utils.hpp for tests/*
