+++ unittests/CMakeLists.txt
@@ -114,7 +114,7 @@ foreach(RUNTIME ${EOSIO_WASM_RUNTIMES})
    set_tests_properties(api_part1_unit_test_${RUNTIME} PROPERTIES COST 5000)
    set_tests_properties(api_part2_unit_test_${RUNTIME} PROPERTIES COST 5000)
    set_tests_properties(api_part3_unit_test_${RUNTIME} PROPERTIES COST 5000)
-   set_tests_properties(api_part4_unit_test_${RUNTIME} PROPERTIES COST 5000)
+   set_tests_properties(checktime_unit_test_${RUNTIME} PROPERTIES COST 5000)
    set_tests_properties(wasm_part1_unit_test_${RUNTIME} PROPERTIES COST 5000)
    set_tests_properties(wasm_part2_unit_test_${RUNTIME} PROPERTIES COST 5000)
    set_tests_properties(wasm_part3_unit_test_${RUNTIME} PROPERTIES COST 5000)
